Moved the elasticsearch connect functions to a central location in the codebase
